# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DayCoach API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task for today
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/complete:
    patch:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as complete or incomplete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteTaskBody"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Partially update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /daily-summary/today/close:
    patch:
      operationId: closeTodaySummary
      tags: [checkin]
      summary: Close out the current day with a short summary and overall status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseDayBody"
      responses:
        "200":
          description: Updated daily summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySummary"
...
  /tasks/validate:
    post:
      operationId: validateTask
      tags: [tasks]
      summary: Validate task specificity and get voice coaching tip if vague
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateTaskBody"
      responses:
        "200":
          description: Validation result with optional coaching tip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateTaskResult"
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        completed:
          type: boolean
        date:
          type: string
          description: ISO date string (YYYY-MM-DD)
        category:
          type: string
          description: One of health, work, learning, mindset
        createdAt:
          type: string
      required:
        - id
        - text
        - completed
        - date
        - category
        - createdAt
    CreateTaskBody:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - text
    CompleteTaskBody:
      type: object
      properties:
        completed:
          type: boolean
      required:
        - completed
    UpdateTaskBody:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
        completed:
          type: boolean
        category:
          type: string
          description: One of health, work, learning, mindset
      minProperties: 1
    CloseDayBody:
      type: object
      properties:
        summaryText:
          type: string
          minLength: 1
          maxLength: 1000
        overallStatus:
          type: string
          enum: [great, partial, missed]
        closureSource:
          type: string
          enum: [voice_agent, manual]
      required:
        - summaryText
        - overallStatus
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    DailySummary:
      type: object
      properties:
        date:
          type: string
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        completionRate:
          type: number
        voicePersonaUsed:
          type: string
        hadCheckin:
          type: boolean
        summaryText:
          type: string
          nullable: true
        overallStatus:
          type: string
          nullable: true
        closedAt:
          type: string
          nullable: true
        closureSource:
          type: string
          nullable: true
      required:
        - date
        - totalTasks
        - completedTasks
        - completionRate
        - voicePersonaUsed
        - hadCheckin
...
    ValidateTaskBody:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - text
    ValidateTaskResult:
      type: object
      properties:
        isVague:
          type: boolean
          description: Whether the task is considered too vague
        suggestion:
          type: string
          description: Coaching suggestion text if vague, null otherwise
          nullable: true
        audioUrl:
          type: string
          description: URL to play coaching audio if vague, null otherwise
          nullable: true
      required:
        - isVague
...