# lib/api-zod/src/generated/types/closeDayBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayCoach API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CloseDayBodyClosureSource } from "./closeDayBodyClosureSource";
import type { CloseDayBodyOverallStatus } from "./closeDayBodyOverallStatus";

export interface CloseDayBody {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  summaryText: string;
  overallStatus: CloseDayBodyOverallStatus;
  closureSource?: CloseDayBodyClosureSource;
}