# lib/api-zod/src/generated/types/dailySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayCoach API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DailySummary {
  date: string;
  totalTasks: number;
  completedTasks: number;
  completionRate: number;
  voicePersonaUsed: string;
  hadCheckin: boolean;
  summaryText?: string | null;
  overallStatus?: string | null;
  closedAt?: string | null;
  closureSource?: string | null;
}