# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DayCoach API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTaskBody {
  /**
   * @minLength 1
   * @maxLength 500
   */
  text?: string;
  completed?: boolean;
  /** One of health, work, learning, mindset */
  category?: string;
}